API path 수정 request, response 수정

+++ src/main/java/com/hhp/concert/controller/payment/dto/PayReservationResponse.java
@@ -9,10 +9,10 @@ public class PayReservationResponse {
 
     private long paymentId;
     private long reservationId;
-    private int paymentAmount;
+    private long paymentAmount;
     private LocalDateTime paymentTime;
 
-    public PayReservationResponse(final long paymentId, final long reservationId, final int paymentAmount, final LocalDateTime paymentTime) {
+    public PayReservationResponse(final long paymentId, final long reservationId, final long paymentAmount, final LocalDateTime paymentTime) {
         this.paymentId = paymentId;
         this.reservationId = reservationId;
         this.paymentAmount = paymentAmount;
